Move generic interfaces into delegate package
