+++ src/main/java/com/alkemy/ong/application/service/DeleteCommentUseCaseService.java
@@ -5,7 +5,7 @@ import com.alkemy.ong.application.exception.ObjectNotFound;
 import com.alkemy.ong.application.exception.OperationNotPermitted;
 import com.alkemy.ong.application.repository.ICommentRepository;
 import com.alkemy.ong.application.service.usecase.IDeleteCommentUseCase;
-import com.alkemy.ong.application.service.usecase.IOperationAllowed;
+import com.alkemy.ong.application.service.delegate.IOperationAllowed;
 import com.alkemy.ong.domain.Identifiable;
 import lombok.AllArgsConstructor;
