+++ src/main/java/com/alkemy/ong/infrastructure/config/spring/security/common/UserAuthentication.java
@@ -2,14 +2,13 @@ package com.alkemy.ong.infrastructure.config.spring.security.common;
 
 import com.alkemy.ong.application.exception.ErrorMessage;
 import com.alkemy.ong.application.exception.InvalidCredentials;
-import com.alkemy.ong.application.service.usecase.IAuthenticateUser;
+import com.alkemy.ong.application.service.delegate.IAuthenticateUser;
 import com.alkemy.ong.domain.User;
 import lombok.AllArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -17,10 +16,6 @@ import org.springframework.stereotype.Component;
 public class UserAuthentication implements IAuthenticateUser {
 
   private AuthenticationManager authenticationManager;
-
-  private UserDetailsService userDetailsService;
-
-  private PasswordEncoder passwordEncoder;
 
   @Override
   public void authenticate(User user) {
